backend/requirements.txt
django==5.1.3
djangorestframework==3.15.2
django-environ==0.11.2
django-allauth==65.0.0
drf-spectacular==0.27.2
httpx==0.27.2
tenacity==9.0.0
psycopg2-binary==2.9.10

pytest==8.3.3
pytest-django==4.9.0


